refactor: move components and tests
